Handling btach payment errors
